+++ src/components/load-button.js
@@ -1,0 +1,3 @@
+const getLoadButton = () => `<button class="load-more" type="button">load more</button>`;
+
+export {getLoadButton as button};
